♻️components - Refactored: organized structure by creating folders
