+++ app/components/Study/StudyCard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../page.module.css";
+import styles from "../../page.module.css";
 
 const StudyCard = ({ title, description, date }) => {
   return (
