+++ app/components/Study/StudyFilteredList.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const StudyFilteredList = () => {
+	return (
+		<div>StudyFilteredList</div>
+	)
+}
+
+export default StudyFilteredList
