+++ app/components/Study/StudyList.jsx
